Create codeql.yml (code scanning workflow)
Currently disabled until some further implementation is done, but I just want to lay some groundwork for future infrastructure. 

This is a basic CodeQL workflow (GitHub Action) that will autobuild (or try to) a repositories code and scan it against a huge database of security vulnerabilities and more. Slap a cronjob on it and we're ballin outta control.